todo/requirements/common.txt
Django~=3.1.1
djangorestframework
djangorestframework-jwt
django-cors-headers
Pillow
mysqlclient-2.0.1
boto3
django-storages